+++ tests/ttnn/unit_tests/test_to_layout.py
@@ -3,10 +3,10 @@
 # SPDX-License-Identifier: Apache-2.0
 
 from loguru import logger
-
-import math
 import pytest
+
 import torch
+
 import ttnn
 
 from tests.ttnn.utils_for_testing import assert_with_pcc, check_with_pcc_without_tensor_printout
@@ -326,140 +326,6 @@ def test_to_layout_page_error(shape, device):
     torch_output = torch_tensor
     assert torch_output.shape == output_tensor.shape
     assert_with_pcc(torch_output, output_tensor, 0.9999)
-
-
-@pytest.mark.parametrize("dtype", [ttnn.bfloat8_b, ttnn.bfloat16])
-@pytest.mark.parametrize("use_pack_untilize", [True, False])
-@pytest.mark.parametrize(
-    "tensor_shape",
-    [
-        [1, 1, 512, 512],
-        [1, 1, 1024, 512],
-        [1, 1, 512, 1024],
-    ],
-)
-@pytest.mark.parametrize(
-    "input_memory_layout",
-    [
-        ttnn.TensorMemoryLayout.HEIGHT_SHARDED,
-        ttnn.TensorMemoryLayout.WIDTH_SHARDED,
-        ttnn.TensorMemoryLayout.BLOCK_SHARDED,
-        ttnn.TensorMemoryLayout.INTERLEAVED,
-    ],
-)
-@pytest.mark.parametrize(
-    "input_shard_orientation",
-    [
-        ttnn.ShardOrientation.ROW_MAJOR,
-        ttnn.ShardOrientation.COL_MAJOR,
-    ],
-)
-@pytest.mark.parametrize(
-    "output_memory_layout",
-    [
-        ttnn.TensorMemoryLayout.HEIGHT_SHARDED,
-        ttnn.TensorMemoryLayout.WIDTH_SHARDED,
-        ttnn.TensorMemoryLayout.BLOCK_SHARDED,
-        ttnn.TensorMemoryLayout.INTERLEAVED,
-    ],
-)
-@pytest.mark.parametrize(
-    "output_shard_orientation",
-    [
-        ttnn.ShardOrientation.ROW_MAJOR,
-        ttnn.ShardOrientation.COL_MAJOR,
-    ],
-)
-@pytest.mark.parametrize(
-    "num_shard_cores, standard_shard_core_grid, block_shard_core_grid",
-    [
-        [
-            4,
-            ttnn.CoreRangeSet({ttnn.CoreRange(ttnn.CoreCoord(0, 0), ttnn.CoreCoord(0, 3))}),
-            ttnn.CoreRangeSet({ttnn.CoreRange(ttnn.CoreCoord(0, 0), ttnn.CoreCoord(1, 1))}),
-        ],
-        [
-            16,
-            ttnn.CoreRangeSet(
-                {
-                    ttnn.CoreRange(ttnn.CoreCoord(0, 0), ttnn.CoreCoord(0, 7)),
-                    ttnn.CoreRange(ttnn.CoreCoord(1, 0), ttnn.CoreCoord(1, 7)),
-                }
-            ),
-            ttnn.CoreRangeSet({ttnn.CoreRange(ttnn.CoreCoord(0, 0), ttnn.CoreCoord(3, 3))}),
-        ],
-    ],
-)
-def test_untilize_single_core(
-    device,
-    dtype,
-    use_pack_untilize,
-    tensor_shape,
-    input_memory_layout,
-    input_shard_orientation,
-    output_memory_layout,
-    output_shard_orientation,
-    num_shard_cores,
-    standard_shard_core_grid,
-    block_shard_core_grid,
-):
-    num_tensor_dims = len(tensor_shape)
-    tensor_height = 1
-    for i in range(num_tensor_dims - 1):
-        tensor_height *= tensor_shape[i]
-    tensor_width = tensor_shape[num_tensor_dims - 1]
-
-    # Shard shapes
-    height_sharded_shard_shape = (tensor_height // num_shard_cores, tensor_width)
-    width_sharded_shard_shape = (tensor_height, tensor_width // num_shard_cores)
-    block_sharded_shard_shape = (
-        tensor_height // int(math.sqrt(num_shard_cores)),
-        tensor_width // int(math.sqrt(num_shard_cores)),
-    )
-
-    # Shard Memory Layout Map
-    shard_memory_layout_map = {
-        ttnn.TensorMemoryLayout.HEIGHT_SHARDED: {
-            "shard_grid": standard_shard_core_grid,
-            "shard_shape": height_sharded_shard_shape,
-        },
-        ttnn.TensorMemoryLayout.WIDTH_SHARDED: {
-            "shard_grid": standard_shard_core_grid,
-            "shard_shape": width_sharded_shard_shape,
-        },
-        ttnn.TensorMemoryLayout.BLOCK_SHARDED: {
-            "shard_grid": block_shard_core_grid,
-            "shard_shape": block_sharded_shard_shape,
-        },
-    }
-
-    # Input memory config
-    input_memory_config = ttnn.MemoryConfig(ttnn.TensorMemoryLayout.INTERLEAVED, ttnn.BufferType.L1)
-    if input_memory_layout != ttnn.TensorMemoryLayout.INTERLEAVED:
-        shard_memory_layout = shard_memory_layout_map[input_memory_layout]
-        input_shard_spec = ttnn.ShardSpec(
-            shard_memory_layout["shard_grid"], shard_memory_layout["shard_shape"], input_shard_orientation
-        )
-        input_memory_config = ttnn.MemoryConfig(input_memory_layout, ttnn.BufferType.L1, input_shard_spec)
-
-    # Output memory config
-    output_memory_config = ttnn.MemoryConfig(ttnn.TensorMemoryLayout.INTERLEAVED, ttnn.BufferType.L1)
-    if output_memory_layout != ttnn.TensorMemoryLayout.INTERLEAVED:
-        shard_memory_layout = shard_memory_layout_map[output_memory_layout]
-        output_shard_spec = ttnn.ShardSpec(
-            shard_memory_layout["shard_grid"], shard_memory_layout["shard_shape"], output_shard_orientation
-        )
-        output_memory_config = ttnn.MemoryConfig(output_memory_layout, ttnn.BufferType.L1, output_shard_spec)
-
-    # Test
-    input_torch_tensor = torch.randn(tensor_shape, dtype=torch.bfloat16)
-    input_ttnn_tensor = ttnn.from_torch(input_torch_tensor, dtype=dtype, layout=ttnn.TILE_LAYOUT)
-    input_ttnn_tensor = ttnn.to_device(input_ttnn_tensor, device, memory_config=input_memory_config)
-    ttnn_output_tensor = ttnn.untilize(
-        input_ttnn_tensor, memory_config=output_memory_config, use_multicore=False, use_pack_untilize=use_pack_untilize
-    )
-
-    assert_with_pcc(input_torch_tensor, ttnn.to_torch(ttnn_output_tensor), 0.9999)
 
 
 @pytest.mark.parametrize("shape", [[64, 7680]])
